Refactor some tests to work better
