+++ src/main/scala/io/epiphanous/flinkrunner/model/SchemaRegistryConfig.scala
@@ -1,18 +1,10 @@
 package io.epiphanous.flinkrunner.model
 
-import com.amazonaws.auth.AWSStaticCredentialsProvider
-import com.amazonaws.services.schemaregistry.common.AWSSchemaRegistryClient
-import com.amazonaws.services.schemaregistry.common.configs.GlueSchemaRegistryConfiguration
 import com.amazonaws.services.schemaregistry.utils.AWSSchemaRegistryConstants
 import com.typesafe.config.{
   ConfigFactory,
   ConfigObject,
   ConfigValueFactory
-}
-import io.confluent.kafka.schemaregistry.client.{
-  CachedSchemaRegistryClient,
-  MockSchemaRegistryClient,
-  SchemaRegistryClient
 }
 import io.confluent.kafka.serializers.KafkaAvroDeserializerConfig
 import io.epiphanous.flinkrunner.model.SchemaRegistryType.{
@@ -21,19 +13,7 @@ import io.epiphanous.flinkrunner.model.SchemaRegistryType.{
 }
 import io.epiphanous.flinkrunner.util.ConfigToProps.RichConfigObject
 import io.epiphanous.flinkrunner.util.StreamUtils.RichProps
-import software.amazon.awssdk.auth.credentials.{
-  AwsBasicCredentials,
-  AwsCredentialsProvider,
-  DefaultCredentialsProvider,
-  StaticCredentialsProvider
-}
-import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration
-import software.amazon.awssdk.core.retry.RetryPolicy
-import software.amazon.awssdk.http.urlconnection.UrlConnectionHttpClient
-import software.amazon.awssdk.regions.Region
-import software.amazon.awssdk.services.glue.GlueClient
 
-import java.net.URI
 import java.util
 import scala.util.{Failure, Success, Try}
 
@@ -45,43 +25,6 @@ case class SchemaRegistryConfig(
     headers: util.Map[String, String] = new util.HashMap()) {
   val url: String =
     props.getOrDefault("schema.registry.url", "na")
-
-  // useful for testing
-  @transient
-  lazy val confluentClient: SchemaRegistryClient =
-    if (url.startsWith("mock")) new MockSchemaRegistryClient()
-    else
-      new CachedSchemaRegistryClient(
-        url,
-        cacheCapacity,
-        props,
-        headers
-      )
-
-  // useful for testing
-  @transient
-  lazy val glueClient: GlueClient = {
-    val accessKeyId         = props.getOrDefault("accessKeyId", "")
-    val secretAccessKey     = props.getOrDefault("secretAccessKey", "")
-    val credentialsProvider =
-      if (accessKeyId.nonEmpty && secretAccessKey.nonEmpty)
-        StaticCredentialsProvider.create(
-          AwsBasicCredentials.create(accessKeyId, secretAccessKey)
-        )
-      else DefaultCredentialsProvider.builder().build()
-    val glueConfig          = new GlueSchemaRegistryConfiguration(props)
-    GlueClient
-      .builder()
-      .credentialsProvider(credentialsProvider)
-      .httpClient(UrlConnectionHttpClient.builder.build)
-      .region(Region.of(glueConfig.getRegion))
-      .endpointOverride(new URI(glueConfig.getEndPoint))
-      .build()
-  }
-
-  @transient
-  lazy val awsSchemaRegistryClient: AWSSchemaRegistryClient =
-    new AWSSchemaRegistryClient(glueClient)
 }
 
 object SchemaRegistryConfig {
