+++ src/test/scala/io/epiphanous/flinkrunner/serde/ConfluentAvroRegistryKafkaRecordSerdeSpec.scala
@@ -2,6 +2,7 @@ package io.epiphanous.flinkrunner.serde
 
 import io.epiphanous.flinkrunner.model._
 import org.apache.flink.api.scala._
+//import org.apache.flink.api.scala._ //<--- optimize imports might remove this
 
 class ConfluentAvroRegistryKafkaRecordSerdeSpec extends SerdeTestFixtures {
 
@@ -14,7 +15,7 @@ class ConfluentAvroRegistryKafkaRecordSerdeSpec extends SerdeTestFixtures {
   property("BWrapper confluent serde round trip scale") {
     val test = SchemaRegistrySerdeTest[BWrapper, BRecord, MyAvroADT]()
     test.init()
-    genPop[BWrapper](10000).zipWithIndex.foreach { case (b, i) =>
+    genPop[BWrapper](1000).zipWithIndex.foreach { case (b, i) =>
       if (i % 100 == 0) println(i)
       test.runTest(b)
     }
